Match the repository usage in both stages of the build.
Seems that in a multi-stage pipeline, the collectionId and the definitionId in each of the stages is shared. Azure uses those to find out which repository to look for a manifest file. When the two stages have different repository lists, it seems that the previous work directory is discarded. Since both stages had different repository lists, one stage was overriding the other all the time. Since GC doesn't happen (because there's no maintenance), the file system was being filled all the time.

In theory.