WiX: add `SDKROOT`, `DEVELOPER_DIR` environment variables
Currently, as we only setup a single SDK which is the build host, set
the SDKROOT environment variable by default. In the future, this should
be conditionalised to if the host system matches the configuration that
is being built here.  The `DEVELOPER_DIR` environment variable provides
a reference point to access XCTest.