swift-format: complete the addition work
Add support for building swift-format via GitHub actions.  This currently releases only an artifact as it requires the Swift 5.5 snapshot to build.  Additionally, it is currently using a custom fork of swift-format as it is awaiting apple/swift-format#253 to repair the build.  Once the necessary changes are merged, it may make sense to release a build.