SwiftInstaller: handle usage before VS installation better
In the case that the WinSDK or VS have not yet been installed, we would
fail due to an invalid buffer usage for the construction of a
`std::filesystem::path`.  This avoids that by ensuring that if we have
no value we do not attempt to create the path.  It also simplifies the
logic and relies more on logging.