Update README.md
Swift 5.2 and 5.3 are quite old at this point.  Remove the status badges for them.  The only state that is of use here is the main branch, and even that remains questionable as swift.org should be preferred for releases.  The Windows builds are the only ones being actively tended to, so remove the other status badges as well.