Update swift-corelibs.yml
Fix the build rules for the corelibs, build nearly a complete SDK.  This is sufficient for getting test coverage for XCTest for the time being.  This change requires apple/swift-corelibs-xctest#332 to be merged for the tests to pass.  dispatch_overcommit has long been flakey on the CI environments but does not flake on real hardware.  The Foundation invalidBundlePath test is known to be currently broken after the last merge for Foundation.