GHA: package the SDK and runtime
This adds the packaging steps for the runtime and SDK.  They are two halves of
the same build: the runtime is the redistributable portion of the SDK, and the
SDK is the developer component.  We package them in a single step as they are
packaging subets of the same files, and it does not seem valuable to spin up a
new environment to package up a different subset of the files.