build: introduce the install root
This introduces a new install root variable as we now need to
restructure the staged installation further.  We will need to
now isolate the runtime components and the SDK much similarly
to the actual installation.  This is the obvious evolution of
the staging as it becomes more similar to the actual installs
which support re-rooting.