CI: adjust the module map deployment
Adjust the CI rules for the changes to the source files.  Ideally this
would be possible to automate without having to do manually, however,
that operation is encoded into the Swift installer Custom Action.  This
is not compatible with older snapshots unfortunately, but hopefully we
will not need to build those again, but it is possible to determine the
source list and conditionalise the behaviour on that if necessary.