GHA: enable additional debug information generation and collection
Enable debug info generation and collection for the standard library and
the devtools.  Refactor some code to avoid duplication of definitions,
hoist the flag computation to a single site.  Selectively extend the
flags to push for a bit more performance (e.g. `/Oi`).  Ensure that we
are always enabling ICF and performing DCE.