GHA: add SwiftFormat as part of the toolchain build
Because we do not have a stable ABI on Windows and the toolchain is
still under rapid development, build and package a version of
SwiftFormat for developers to install locally for productivity.  This is
a divergence from the upstream toolchain.  Ideally, this will be removed
and subsumed by the nicklockwood/SwiftFormat repository.