build: disable debug information on release builds
Disable the default build from emitting debug information.  This will
speed up the build and reduce the output size.  This is particularly
important for the Swift side which will emit a large amount of metadata
into the final binaries which is not stripped.

This is intended to be the default build type for CI.  For local builds,
developers may use `-BuildType RelWithDebInfo` to get debug information
emitted.