Delete Windows.md
The majority of this is no longer relevant.  It is preferred to access the nightlies through swift.org.  The installation is no longer relevant as the installer has evolved to automate the process here and the module maps no longer need to be copied over.  ICU is statically linked, SPM is supported.  The installation structure has changed and the only piece that remains relevant here is the distribution aspect which is also in the process of being revamped.  This guide is entirely unhelpful in the light of all these changes.