[gha][binary sizes] Fix inputs when using workflow_call
Previously this workflow read the toolchain_version input using
`github.event.inputs` which carries the same `event` values as the calling
workflow This means `github.event.inputs` is empty when the workflow is
triggered via workflow_call and the calling worfklow was triggered as part of
a `schedule` event. This was not caught during testing because we triggered
the calling worfklow using the `pull_request` event instead. This change
switches to using the `input` context which is always set to the correct
value regardless if whether our workflow is triggered via
`pull_request`, `workflow_call` or `workflow_dispatch`.