Use toolchain clang to build compilers
Clang 18 is the default version of Clang on Github runner images, and has a broken definition of the __prefetch intrinsic on arm64 which causes build failures. We can instead use the Clang that comes with the pinned toolchain (Clang 16 on Swift 5.x releases), which does not have this broken definition.