GHA: introduce a new macros build phase
Introduce a new macros build phase for the Foundation (and eventually
Testing) macros. This depends on the previously introduced stdlib phase
as the macros use the standard library. We build the macros for the
compiler hosts, reusing them to build the SDK content. Wire this into
the packaging step for the build tools as well.

Co-authored-by: Alex Lorenz <arphaman@gmail.com>