[stdlib] Change paths for stdlib job and artifacts (#888)
Currently, in the `stdlib` job, the toolchain is extracted to
`BuildRoot/Library`, and the Swift Standard Library is also installed to
`BuildRoot/Library`. Other jobs typically extract the toolchain to
`BinaryCache/Library`. This changes the extraction path to
`BinaryCache/Library` in the `stdlib` job, bringing it in-line with the
rest of the build.

In addition, this changes the Swift Standard Library artifact root to be
`BuildRoot/Library`, rather than the full path to the platform SDK,
`BuildRoot/Library/Developer/Platforms/${{ matrix.os }}.platform`. This
is in preparation for adapting the job to build for macOS, where the
Swift Standard Library will be installed under
`Library/Toolchains/unknown-Asserts-development.xctoolchain/usr`.

Extracted from #843