Temporarily specify Windows SDK and MSVC versions
This is necessary to work around an issue with the latest version
(26100) of the Windows SDK where changes in headers result in circular
module dependencies.
For the early swift-driver, this bumps the bootstrap toolchain to use
Swift 6.1 to work around MSVC 14.43 requiring Clang 18+.
For the compilers build, this installs and uses MSVC 14.42 because the
current build does not work with MSVC 14.43, either with Swift 6.0 or
Swift 6.1.

See #909 for details.