[Take 2] Re-enable toolchain-binary-sizes workflow (#286)
In my previous change (#284), I added triggering on push/tags. I was wrong. that too does not work (see https://github.com/orgs/community/discussions/27028). seems like there is an embargo on any action that is triggrerd by automation to trigger another workflow. we have had a successful release (20250527.3) and it did not trigger the workflow.

This change calls the binary size workflow from build-toolchain.yml instead.

I now need to cherry pick this change, and the previous changes to be in release/6.1 and release/6.2 as well.