Add reusable action for MSVC + Win SDK overrides (#956)
This introduces a new reusable action, `setup-build`, which will
eventually replace all of the custom installation steps in the
swift-build jobs. Currently it only supports providing explicit versions
for MSVC and the Windows SDK.

This also adds tests for the `setup-build` action, with a configurable
runner, ensuring we can catch issues early as runner images get updated.