actions: Add a configure-cmake-project action (#999)
This brings in a modified version of `Build-CMakeProject` in `build.ps1` from
the swift repository as a custom action. This reduces the differences in the
build setup between the swift repository and this repository, making it easier
to keep both builds in sync.

The new reusable action only configures the CMake project in order to keep the
build and install steps as separate steps in the GitHub Actions UI.